Move functions from util to static methods, increase tolerance for test
