+++ src/pydvl/influence/torch/util.py
@@ -52,8 +52,6 @@ __all__ = [
     "torch_dataset_to_dask_array",
     "EkfacRepresentation",
     "empirical_cross_entropy_loss_fn",
-    "rank_one_mvp",
-    "inverse_rank_one_update",
     "LossType",
     "ModelParameterDictBuilder",
     "BlockMode",
@@ -610,135 +608,6 @@ class TorchLinalgEighException(Exception):
             f" Inspect the original exception message: \n{str(original_exception)}"
         )
         super().__init__(err_msg)
-
-
-def rank_one_mvp(x: torch.Tensor, v: torch.Tensor) -> torch.Tensor:
-    r"""
-    Computes the matrix-vector product of xx^T and v for each row in X and V without
-    forming xx^T and sums the result. Here, X and V are matrices where each row
-    represents an individual vector. Effectively it is computing
-
-    $$ V@( \frac{1}{N}\sum_i^N x[i]x[i]^T) $$
-
-    Args:
-        x: Matrix of vectors of size `(N, M)`.
-        v: Matrix of vectors of size `(B, M)` to be multiplied by the corresponding
-            $xx^T$.
-
-    Returns:
-        A matrix of size `(B, N)` where each column is the result of xx^T v for
-            corresponding rows in x and v.
-    """
-    if v.ndim == 1:
-        result = torch.einsum("ij,kj->ki", x, v.unsqueeze(0)) @ x
-        return result.squeeze() / x.shape[0]
-    return (torch.einsum("ij,kj->ki", x, v) @ x) / x.shape[0]
-
-
-def generate_rank_one_mvp(
-    x: List[torch.Tensor], v: List[torch.Tensor]
-) -> Generator[torch.Tensor, None, None]:
-    x_v_iterator = zip(x, v)
-    x_, v_ = next(x_v_iterator)
-
-    nominator = torch.einsum("i..., k...->ki", x_, v_)
-
-    for x_, v_ in x_v_iterator:
-        nominator += torch.einsum("i..., k...->ki", x_, v_)
-
-    for x_, v_ in zip(x, v):
-        yield torch.einsum("ji, i... -> j...", nominator, x_) / x_.shape[0]
-
-
-def inverse_rank_one_update(
-    x: torch.Tensor, v: torch.Tensor, regularization: float
-) -> torch.Tensor:
-    r"""
-    Performs an inverse-rank one update on x and v. More precisely, it computes
-
-    $$ \sum_{i=1}^n \left(x[i]x[i]^t+\lambda \operatorname{I}\right)^{-1}v $$
-
-    where $\operatorname{I}$ is the identity matrix and $\lambda$ is positive
-    regularization parameter. The inverse matrices are not calculated explicitly,
-    but instead a vectorized version of the
-    [Sherman–Morrison formula](
-    https://en.wikipedia.org/wiki/Sherman%E2%80%93Morrison_formula)
-    is applied.
-
-    Args:
-        x: Input matrix used for the rank one expressions. First dimension is
-            assumed to be the batch dimension.
-        v: Matrix to multiply with. First dimension is
-            assumed to be the batch dimension.
-        regularization: Regularization parameter to make the rank-one expressions
-            invertible, must be positive.
-
-    Returns:
-        Matrix of size $(D, M)$ for x having shape $(N, D)$ and v having shape $(M, D)$.
-    """
-    nominator = torch.einsum("ij,kj->ki", x, v)
-    denominator = x.shape[0] * (regularization + torch.sum(x**2, dim=1))
-    return (v - (nominator / denominator) @ x) / regularization
-
-
-def generate_inverse_rank_one_updates(
-    x: List[torch.Tensor], v: List[torch.Tensor], regularization: float
-) -> Generator[torch.Tensor, None, None]:
-    def _check_batch_dim(t_x, t_v, idx: int):
-        if t_x.ndim <= 1:
-            raise ValueError(
-                f"Provided tensors in the lists must have at least "
-                f"2 dimensions, "
-                f"but found {t_x.ndim=} at {idx=} in list x"
-            )
-
-        if v_.ndim <= 1:
-            raise ValueError(
-                f"Provided tensors in the lists must have at least "
-                f"2 dimensions, "
-                f"but found shape {t_v.ndim=} at {idx=} in list v"
-            )
-
-    def _create_dim_error(x_shape, v_shape, idx: int):
-        return ValueError(
-            f"Entries in the tensor lists must have the same "
-            f"(excluding the first batch dimensions), "
-            f"but found shapes {x_shape} and {v_shape}"
-            f"at {idx=}"
-        )
-
-    if not len(x) == len(v):
-        raise ValueError(
-            f"Provided tensor lists must have the same length, but got"
-            f"{len(x)=} and {len(v)=}"
-        )
-
-    x_v_iterator = enumerate(zip(x, v))
-    index, (x_, v_) = next(x_v_iterator)
-
-    _check_batch_dim(x_, v_, index)
-
-    if x_.shape[1:] != v_.shape[1:]:
-        raise _create_dim_error(x_.shape[1:], v_.shape[1:], index)
-
-    denominator = regularization + torch.sum(x_.view(x_.shape[0], -1) ** 2, dim=1)
-    nominator = torch.einsum("i..., k...->ki", x_, v_)
-    num_data_points = x_.shape[0]
-
-    for k, (x_, v_) in x_v_iterator:
-        _check_batch_dim(x_, v_, k)
-        if x_.shape[1:] != v_.shape[1:]:
-            raise _create_dim_error(x_.shape[1:], v_.shape[1:], k)
-
-        nominator += torch.einsum("i..., k...->ki", x_, v_)
-        denominator += torch.sum(x_.view(x_.shape[0], -1) ** 2, dim=1)
-
-    denominator = num_data_points * denominator
-
-    for x_, v_ in zip(x, v):
-        yield (
-            v_ - torch.einsum("ji, i... -> j...", nominator / denominator, x_)
-        ) / regularization
 
 
 LossType = Callable[[torch.Tensor, torch.Tensor], torch.Tensor]
